refactor 5 move load config to utils
