--- experimental/config/__init__.py
@@ -1,24 +1,0 @@
-# Copyright (c) 2026 verl-project authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Monitor config loaders."""
-
-from __future__ import annotations
-
-from .config import load_monitor_config, load_server_config_file
-
-__all__ = [
-    "load_monitor_config",
-    "load_server_config_file",
-]
