+++ experimental/utils/__init__.py
@@ -23,6 +23,7 @@ from .constants import (
     MonitorRayActor,
     PrometheusScrape,
 )
+from .monitor_config_loader import load_monitor_config, load_server_config_file
 from .opentelemetry_utils import (
     OpenTelemetryTraceCollector,
     resolve_otlp_traces_endpoint,
@@ -44,6 +45,8 @@ __all__ = [
     "MonitorRayActor",
     "OpenTelemetryTraceCollector",
     "PrometheusScrape",
+    "load_monitor_config",
+    "load_server_config_file",
     "merge_labels",
     "resolve_otlp_traces_endpoint",
     "start_metrics_http_server",
